Refactor how federationsender gets created
